Implements material of a test task.
It contains a view and interfaces of model and controller.

The model part maintains a list of element and it will trigger events
if the data was changed.

The controller part has some method for view to call and will call
model's method to change the data.

The view part listens model's events and calls the controller's method
if it cought events from the user.